# Further efficient-ize ApnsPayloadBuilder