--- a/pushy/src/main/java/com/relayrides/pushy/apns/util/ApnsPayloadBuilder.java
+++ b/pushy/src/main/java/com/relayrides/pushy/apns/util/ApnsPayloadBuilder.java
@@ -21,7 +21,7 @@
 package com.relayrides.pushy.apns.util;
 
 import java.io.CharArrayWriter;
-import java.nio.charset.Charset;
+import java.nio.charset.StandardCharsets;
 import java.util.Arrays;
 import java.util.HashMap;
 import java.util.Map;
@@ -77,7 +77,6 @@ public class ApnsPayloadBuilder {
 
     private static final int DEFAULT_PAYLOAD_SIZE = 4096;
 
-    private static final Charset UTF8 = Charset.forName("UTF-8");
     private static final String ABBREVIATION_SUBSTRING = "…";
 
     private static final Gson gson = new GsonBuilder().serializeNulls().create();
@@ -408,7 +407,7 @@ public class ApnsPayloadBuilder {
         gson.toJson(payload, this.buffer);
 
         final String payloadString = this.buffer.toString();
-        final int initialPayloadSize = payloadString.getBytes(UTF8).length;
+        final int initialPayloadSize = payloadString.getBytes(StandardCharsets.UTF_8).length;
 
         final String fittedPayloadString;
 
@@ -421,46 +420,19 @@ public class ApnsPayloadBuilder {
                 this.buffer.reset();
                 gson.toJson(payload, this.buffer);
 
-                final int payloadSizeWithEmptyMessage = this.buffer.toString().getBytes(UTF8).length;
+                final int payloadSizeWithEmptyMessage = this.buffer.toString().getBytes(StandardCharsets.UTF_8).length;
 
                 if (payloadSizeWithEmptyMessage >= maximumPayloadSize) {
                     throw new IllegalArgumentException("Payload exceeds maximum size even with an empty message body.");
                 }
 
-                // We add 2 here because the escaped string will include opening and closing '"' characters that are
-                // already accounted for in payloadSizeWithEmptyMessage.
-                final int maximumEscapedMessageBodySize = maximumPayloadSize - payloadSizeWithEmptyMessage + 2;
+                final int maximumEscapedMessageBodySize = maximumPayloadSize - payloadSizeWithEmptyMessage -
+                        ABBREVIATION_SUBSTRING.getBytes(StandardCharsets.UTF_8).length;
 
-                // Characters in the message body may wind up representing different numbers of bytes as an escaped
-                // JSON string. Assuming a best case of one byte per character, we have a ceiling on the maximum number
-                // of characters we could possibly fit into the message body. We can also figure out a minimum number
-                // of characters; in the worst case (a unicode control character), a single character could expand to
-                // take up six bytes in the escaped JSON string.
-                int left = maximumEscapedMessageBodySize / 6;
-                int right = maximumEscapedMessageBodySize;
+                final String fittedMessageBody = this.alertBody.substring(0,
+                        this.getLengthOfJsonEscapedUtf8StringFittingSize(this.alertBody, maximumEscapedMessageBodySize));
 
-                String fittedMessageBody = null;
-
-                while (left < right) {
-                    final int middle = (left + right) / 2;
-
-                    fittedMessageBody = this.abbreviateString(this.alertBody, middle);
-
-                    this.buffer.reset();
-                    gson.toJson(fittedMessageBody, this.buffer);
-
-                    final int escapedMessageBodySize = this.buffer.toString().getBytes(UTF8).length;
-
-                    if (escapedMessageBodySize == maximumEscapedMessageBodySize) {
-                        break;
-                    } else if (escapedMessageBodySize < maximumEscapedMessageBodySize) {
-                        left = middle + 1;
-                    } else {
-                        right = middle - 1;
-                    }
-                }
-
-                this.replaceMessageBody(payload, fittedMessageBody);
+                this.replaceMessageBody(payload, fittedMessageBody + ABBREVIATION_SUBSTRING);
 
                 this.buffer.reset();
                 gson.toJson(payload, this.buffer);
@@ -497,22 +469,6 @@ public class ApnsPayloadBuilder {
         } else {
             throw new IllegalArgumentException("Payload has no message body.");
         }
-    }
-
-    private String abbreviateString(final String string, final int maximumLength) {
-        final String abbreviatedString;
-
-        if (string.length() <= maximumLength) {
-            abbreviatedString = string;
-        } else {
-            if (maximumLength <= ABBREVIATION_SUBSTRING.length()) {
-                throw new IllegalArgumentException("String is too short to abbreviate.");
-            }
-
-            abbreviatedString = string.substring(0, maximumLength - ABBREVIATION_SUBSTRING.length()) + ABBREVIATION_SUBSTRING;
-        }
-
-        return abbreviatedString;
     }
 
     private Object createAlertObject() {
@@ -593,5 +549,51 @@ public class ApnsPayloadBuilder {
                 && this.localizedActionButtonKey == null && this.alertTitle == null
                 && this.localizedAlertTitleKey == null && this.localizedAlertKey == null
                 && this.localizedAlertArguments == null && this.localizedAlertTitleArguments == null;
+    }
+
+    private int getLengthOfJsonEscapedUtf8StringFittingSize(final String string, final int maximumSize) {
+        int i = 0;
+        int cumulativeSize = 0;
+
+        while (i < string.length()) {
+            final char c = string.charAt(i);
+
+            final int charSize;
+            final int step;
+
+            if (c == '"' || c == '\\' || c == '/' || c == '\b' || c == '\f' || c == '\n' || c == '\r' || c == '\t') {
+                // Character is backslash-escaped in JSON
+                charSize = 2;
+                step = 1;
+            } else if (Character.isISOControl(c)) {
+                charSize = 6;
+                step = 1;
+            } else {
+                // The character will be represented as an un-escaped UTF8 character
+                if (c <= 0x007F) {
+                    charSize = 1;
+                    step = 1;
+                } else if (c <= 0x07FF) {
+                    charSize = 2;
+                    step = 1;
+                } else if (Character.isHighSurrogate(c)) {
+                    charSize = 4;
+                    step = 2;
+                } else {
+                    charSize = 3;
+                    step = 1;
+                }
+            }
+
+            if (cumulativeSize + charSize > maximumSize) {
+                // The next character would put us over the edge; bail out here.
+                break;
+            } else {
+                i += step;
+                cumulativeSize += charSize;
+            }
+        }
+
+        return i;
     }
 }