This does two things to make `ApnsPayloadBuilder` a bit more efficient:

1. It introduces a reusable buffer for JSON generation, which should cut down on buffer allocations and reduce GC pressure when reusing `ApnsPayloadBuilder` instances.
2. It totally changes the algorithm for finding alert bodies that fit within a limited payload size. Previously, we were doing a binary search. In c9c9b89, we just count bytes to figure out how much of the alert body we can include.

Benchmarks:

```
Before:

(messageBodyLength)      (unicodeBlockName)   Mode  Cnt       Score      Error  Units
                512             BASIC_LATIN  thrpt   80  222343.928 ± 4858.257  ops/s
                512  CJK_UNIFIED_IDEOGRAPHS  thrpt   80  305840.679 ± 1133.963  ops/s
               8192             BASIC_LATIN  thrpt   80    3323.071 ±  238.083  ops/s
               8192  CJK_UNIFIED_IDEOGRAPHS  thrpt   80    6939.819 ±   25.177  ops/s

After:

(messageBodyLength)      (unicodeBlockName)   Mode  Cnt       Score      Error  Units
                512             BASIC_LATIN  thrpt   80  222640.631 ± 1577.262  ops/s
                512  CJK_UNIFIED_IDEOGRAPHS  thrpt   80  331607.892 ± 1820.027  ops/s
               8192             BASIC_LATIN  thrpt   80    9927.099 ±   69.721  ops/s
               8192  CJK_UNIFIED_IDEOGRAPHS  thrpt   80   21185.339 ±   83.509  ops/s
```